SID_Plugin.pbi
Added: more functions -- still empty buffer when rending.

Bugs:
Inexplicable Crash:
[17:58:21] [ERROR] SID_Plugin.pbi (Line: 343)
[17:58:21] [ERROR] Invalid memory access. (read error at address 66590218)

No idea why this is happening as it's the last line of code and doesn't direct me to an actual line of code that is a fault, also it's intermittent and cannot be repeated by just running it once, although it may fire on the first attempt or it might not, which makes it nearly impossible to nail down.